Remove Chief Technology Officer title
It has been suggested to me by a recruiter that the CTO title, while
technically true, detracts from my CV, as it gives the impression more
of my time was spent in management tasks than was actually the case.
As really the only technical officer, I did oversee technology and data
efforts and was the point person for outside technology consultants and
service providers. But by far my primary focus was on systems
administratiron and engineering, which is where my interests are.

It also suggests I would be more interested in management roles than I
actually am.